Move pipelines to kfp_ray subdir.
