--- transforms/universal/noop/ray/kfp/v1/Makefile
@@ -1,34 +1,0 @@
-REPOROOT=${CURDIR}/../../../../../..
-WORKFLOW_VENV_ACTIVATE=${REPOROOT}/transforms/venv/bin/activate
-include $(REPOROOT)/transforms/.make.transforms_workflows
-
-SRC_DIR=${CURDIR}/../../
-
-PYTHON_WF := $(shell find ./ -name '*_wf.py')
-YAML_WF := $(patsubst %.py, %.yaml, ${PYTHON_WF})
-
-workflow-venv: .check_python_version ${WORKFLOW_VENV_ACTIVATE}
-
-.PHONY: workflow-build
-workflow-build: workflow-venv
-	@for file in $(YAML_WF); do \
-		$(MAKE) $$file; \
-	done
-
-.PHONY: workflow-test
-workflow-test: workflow-build
-	@for file in $(YAML_WF); do \
-		$(MAKE) .transforms_workflows.test-pipeline TRANSFORM_SRC=${SRC_DIR} PIPELINE_FILE=$$file; \
-	done
-
-.PHONY: workflow-upload
-workflow-upload: workflow-build
-	@for file in $(YAML_WF); do \
-		$(MAKE) .transforms_workflows.upload-pipeline PIPELINE_FILE=$$file; \
-	done
-
-.PHONY: workflow-reconcile-requirements
-workflow-reconcile-requirements:
-	@for file in $(PYTHON_WF); do \
-		$(MAKE) .transforms_workflows.reconcile-requirements PIPELINE_FILE=$$file; \
-	done
